backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
sse-starlette==2.1.3
requests==2.32.3
beautifulsoup4==4.12.3
